Handle stray .mp4.h264 files left from failed MP4Box or no space.
In pikrellcam, if h264 file size is zero, delete it and don't run MP4Box on it.

In web page when deleting a thumb, check for a stray .mp4.h264 which can still
happen if we run out of space after starting the MP4Box.

Add deleting any .mp4.h264 files to _fix_thumbs.

Note, this does not affect fname.h264 files which can be from user opting
for h264 files (unsupported) in pikrellcam.conf (these won't have .mp4.h264).